updated layout code
