+++ src/redux/actions/layout/index.js
@@ -1,0 +1,7 @@
+import addWindow from 'redux/actions/layout/addWindow';
+import updateLayout from 'redux/actions/layout/updateLayout';
+
+export {
+  addWindow,
+  updateLayout,
+};
