+++ src/redux/actions/layout/updateLayout.js
@@ -1,0 +1,11 @@
+import { UPDATE_LAYOUT } from 'redux/actionTypes/layout';
+
+const updateLayout = (windows, panel) => (dispatch) => dispatch({
+  type: UPDATE_LAYOUT,
+  payload: {
+    windows,
+    panel,
+  },
+});
+
+export default updateLayout;
